Split Volt test runtime modules
