+++ guides/rfcs/testing.md
@@ -156,12 +156,12 @@ Proposed modules:
 
 Runtime assets and types:
 
-- `priv/ts/test/core.ts` — `describe`, `test`, `test.skip`, `test.todo`, hooks, context `skip`, and `expect`. Initial implementation provides globals and structured result serialization, dogfooded by `test/volt/test/fixtures/core_api.test.ts` importing a relative fixture module.
+- `priv/ts/test/core.ts` — runtime entry that installs the `volt:test` globals. The implementation is split into focused modules for API registration, matcher behavior, state, errors, formatting, and runner protocol. Initial implementation provides `describe`, `describe.skip`, `describe.todo`, `test`, `test.skip`, `test.todo`, hooks, context `skip`, and `expect`, dogfooded by `test/volt/test/fixtures/core_api.test.ts` importing a relative fixture module.
 - `priv/types/test/volt-test.d.ts` — canonical `volt:test` type contract, including the shared `Volt.Test` namespace used by the runtime implementation so the API shape is not duplicated between implementation and ambient module declarations.
 - `priv/types/internal/support-globals.d.ts` — private Volt support-module globals such as `$css` and `$mod_url`.
 - `priv/types/client/hmr.d.ts` — app-facing HMR browser globals.
 - `priv/types/frameworks/*.d.ts` — framework/compiler-specific declaration shims split per framework.
-- `priv/ts/test/core.ts` — collection/execution protocol used by `Volt.Test.Runner`.
+- `priv/ts/test/runner.ts` — collection/execution protocol used by `Volt.Test.Runner` and imported by the `core.ts` entry.
 - Future `volt:test/browser` runtime assets — browser fixture API and JS proxies for Playwright commands.
 
 ### Execution model
@@ -210,7 +210,8 @@ Suggested runtime options:
 ```elixir
 Volt.JS.Runtime.ensure!(
   name: {:global, {:volt_test_runtime, profile}},
-  entry: {:volt_asset, "test/core.ts"},
+  entry: {:external_path, Volt.Priv.path({:volt, "ts"}, "test/core.ts")},
+  bundle: true,
   bundle: true,
   apis: [:browser, :node],
   handlers: Volt.Test.Runner.handlers(context)
@@ -316,7 +317,7 @@ URL: http://localhost:4002/
 
 - Add config/discovery.
 - Add `Volt.Test.ExUnit.install/1` for `test/test_helper.exs`.
-- Add `priv/ts/test/core.ts` with `describe`, `test`, hooks, and core `expect` matchers. Implemented and dogfooded by `test/volt/test/fixtures/core_api.test.ts`.
+- Add the `priv/ts/test/core.ts` entry and focused runtime modules with `describe`, `test`, hooks, and core `expect` matchers. Implemented and dogfooded by `test/volt/test/fixtures/core_api.test.ts`.
 - Run one ExUnit test per JS/TS file. Initial implementation is in place via generated bridge modules.
 - Support async tests/promises.
 - Add tests for pass/fail/async/hook behavior.
