+++ priv/ts/test/format.ts
@@ -1,0 +1,38 @@
+export function format(value: unknown) {
+  if (typeof value === 'string') return JSON.stringify(value)
+
+  try {
+    return JSON.stringify(value)
+  } catch {
+    return String(value)
+  }
+}
+
+export function deepEqual(left: unknown, right: unknown): boolean {
+  if (Object.is(left, right)) return true
+
+  if (typeof left !== typeof right) return false
+
+  if (typeof left !== 'object' || left === null || right === null) {
+    return false
+  }
+
+  if (Array.isArray(left) || Array.isArray(right)) {
+    if (!Array.isArray(left) || !Array.isArray(right)) return false
+    if (left.length !== right.length) return false
+    return left.every((value, index) => deepEqual(value, right[index]))
+  }
+
+  const leftRecord = left as Record<string, unknown>
+  const rightRecord = right as Record<string, unknown>
+  const leftKeys = Object.keys(leftRecord)
+  const rightKeys = Object.keys(rightRecord)
+
+  if (leftKeys.length !== rightKeys.length) return false
+
+  return leftKeys.every(
+    (key) =>
+      Object.prototype.hasOwnProperty.call(rightRecord, key) &&
+      deepEqual(leftRecord[key], rightRecord[key])
+  )
+}
